refactor: rename DocumentDetail to DocumentDetails
