+++ components/lookup/document/DocumentDetails.tsx
